requirements.txt
Flask
Flask-Login
openpyxl
reportlab
Werkzeug
python-dotenv
